multi_agent_weather/pyproject.toml
[project]
name = "multi-agent-weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.123.10",
    "google-adk[a2a]>=1.21.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "requests>=2.32.3",
    "google-generativeai>=0.8.3",
]
